test: updated tsconfig and jest config for test using ESM from did-jwt library
