+++ .gitignore
@@ -5,3 +5,4 @@ build
 site
 .yarn
 .pnp.*
+database.sqlite
