+++ package.json
@@ -39,16 +39,14 @@
   "scripts": {
     "start": "microbundle watch  -i ./src/sdk/index.ts -o build/sdk/index.js",
     "build:sdk": "microbundle --compress=false -i ./src/sdk/index.ts -o build/sdk/index.js",
-    "build:util": "microbundle --compress=false -i ./src/sdk/util/index.ts -o build/util/index.js --tsconfig ./tsconfig.util.json",
     "build:cli": "microbundle --compress=false --target=node -f cjs -i ./src/cli/index.ts -o build/cli/index.js",
-    "build:api": "microbundle --compress=false -i ./src/api/index.ts -o build/api/index.js --tsconfig ./tsconfig.api.json && cp ./src/api/tonomy.css ./build/api/tonomy.css",
-    "build": "yarn run build:sdk && yarn run build:cli && yarn run build:api && yarn run build:util",
+    "build": "yarn run build:sdk && yarn run build:cli",
     "test:setup": "./test/build-and-run-services.sh start",
     "test:setup-down": "./test/build-and-run-services.sh stop",
-    "test:unit": "jest --selectProjects \"Unit tests\"",
-    "test:integration": "jest --detectOpenHandles --selectProjects \"Integration tests\"",
-    "test:governance": "jest --detectOpenHandles --selectProjects \"Governance tests\"",
-    "test": "jest --detectOpenHandles",
+    "test:unit": "yarn run test --selectProjects \"Unit tests\"",
+    "test:integration": "yarn run test --selectProjects \"Integration tests\"",
+    "test:governance": "yarn run test --selectProjects \"Governance tests\"",
+    "test": "cross-env NODE_OPTIONS=\"--experimental-vm-modules\" jest --detectOpenHandles",
     "cli": "node ./build/cli/index.cjs",
     "lint": "eslint src test --ext .ts --resolve-plugins-relative-to .",
     "lint:fix": "yarn run lint -- --fix",
@@ -73,6 +71,7 @@
     "@typescript-eslint/parser": "^7.13.1",
     "address": "^2.0.2",
     "babel-jest": "^29.7.0",
+    "cross-env": "^7.0.3",
     "eslint": "^9.5.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jest": "^28.6.0",
