--- src/api/index.ts
@@ -1,7 +1,0 @@
-import { ExternalUser } from './externalUser';
-import { setSettings } from '../sdk';
-
-export const api = {
-    ExternalUser,
-    setSettings,
-};
