--- test/jest.setup.ts
@@ -1,3 +1,0 @@
-import { setTestSettings } from './helpers/settings';
-
-setTestSettings();
